+++ Sources/Tachikoma/Providers/ProviderParser.swift
@@ -328,6 +328,8 @@ public enum ProviderParser {
             .minimax(.m27)
         case "minimax-m2.7-highspeed", "minimax-m2-7-highspeed", "m2.7-highspeed", "m2-7-highspeed":
             .minimax(.m27Highspeed)
+        case "minimax-m3", "m3":
+            .minimax(.m3)
         default:
             nil
         }
@@ -339,6 +341,8 @@ public enum ProviderParser {
             .minimaxCN(.m27)
         case "minimax-m2.7-highspeed", "minimax-m2-7-highspeed", "m2.7-highspeed", "m2-7-highspeed":
             .minimaxCN(.m27Highspeed)
+        case "minimax-m3", "m3":
+            .minimaxCN(.m3)
         default:
             nil
         }